event resolver refactored
