+++ src/graphql/resolvers/event/createEvent.js
@@ -1,20 +1,9 @@
-import Event from '../../mongoose/models/event'
-import User from '../../mongoose/models/user'
-import { transformEvent } from './merge'
-import { isAuth } from '../../helpers/isAuth'  
+import Event from '../../../mongoose/models/event'
+import User from '../../../mongoose/models/user'
+import { transformEvent } from '../merge'
+import { isAuth } from '../../../helpers/isAuth'  
 
-export const eventResolver = {
-  events: async () => {
-    try {
-      const events = await Event.find()
-      return events.map(event => {
-        return transformEvent(event)
-      })
-    } catch (err) {
-      throw err
-    }
-  },
-
+export const createEvent ={
   createEvent: async (args, req) => {
     isAuth(req.isAuth)
     const event = new Event({
