+++ src/graphql/resolvers/event/index.js
@@ -1,0 +1,7 @@
+import { events } from './events'
+import { createEvent } from './createEvent'
+
+export const eventResolver = {
+  ...events,
+  ...createEvent
+}
